apps/backend/pyproject.toml
[project]
name = "shelfy-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "supabase>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pdfplumber>=0.11.4",
    "pypdf>=5.1.0",
    "chromadb>=0.5.20",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
